+++ src/Tabs/Tabs.js
@@ -5,18 +5,26 @@ import PropTypes from 'prop-types';
 import WixComponent from '../BaseComponents/WixComponent';
 import SideContent from './core/SideContent';
 import TabItems from './core/TabItems';
+import classNames from 'classnames';
 import * as TabPropTypes from './core/constants/tab-prop-types';
 import styles from './Tabs.scss';
 
 
 class Tabs extends WixComponent {
 
+  static defaultProps = {
+    hasDivider: true
+  }
+
   render() {
-    const {sideContent} = this.props;
+    const {sideContent, hasDivider} = this.props;
     const tabItemsProps = omit(['sideContent'], this.props);
+    const className = classNames(styles.container, {
+      [styles.hasDivider]: hasDivider
+    });
 
     return (
-      <div className={styles.container}>
+      <div className={className}>
         <TabItems {...tabItemsProps}/>
         <SideContent content={sideContent}/>
       </div>
@@ -32,7 +40,8 @@ Tabs.propTypes = {
   type: TabPropTypes.type,
   width: TabPropTypes.width,
   sideContent: TabPropTypes.sideContent,
-  dataHook: PropTypes.string
+  dataHook: PropTypes.string,
+  hasDivider: PropTypes.bool
 };
 
 export default Tabs;